cv subscription

+++ src/cv/cv.resolver.ts
@@ -1,10 +1,13 @@
-import { Resolver, Query, Args, Context, Mutation } from '@nestjs/graphql';
+import { Resolver, Query, Args, Context, Mutation, Subscription } from '@nestjs/graphql';
 import { CV } from './cv.entity';
 import { CreateCVInput } from './inputs/create-cv.input';
 import { UpdateCVInput } from './inputs/update-cv.input';
 import { CVService } from './cv.service';
 import { UserService } from "@/user/user.service";
 import { SkillService } from "@/skill/skill.service";
+import { Inject } from '@nestjs/common';
+import { PubSub } from 'graphql-yoga';
+import { Skill } from '@/skill/skill.entity';
 
 @Resolver(() => CV)
 export class CvResolver {
@@ -13,11 +16,12 @@ export class CvResolver {
     private readonly cvService: CVService, 
     private readonly userService : UserService,
     private readonly skillService : SkillService,
+    @Inject('PUB_SUB') private pubSub: PubSub<any>
   ) {}
 
   @Query(() => [CV])
   async cvs() {
-    return this.cvService.findAll().map(cv => this.hydrate(cv));
+    return this.cvService.findAll().map(async cv => await this.hydrate(cv));
   }
 
   @Query(() => CV, { nullable: true })
@@ -25,14 +29,13 @@ export class CvResolver {
     @Args('id') id: number
   ) {
     const cv = this.cvService.findOne(id);
-    return cv ? this.hydrate(cv) : null;
+    return cv ? await this.hydrate(cv) : null;
   }
 
   @Mutation(() => CV)
   async createCV(
     @Args('input') input: CreateCVInput
   ) {
-    // check if userId and skillIds are valid
     input.userId = parseInt(input.userId.toString(), 10);
     const user = this.userService.findOne(input.userId);
     if (!user) {
@@ -46,14 +49,15 @@ export class CvResolver {
     }
 
     const rawCv = this.cvService.create(input);
-    return this.hydrate(rawCv);
+    const hydrated = await this.hydrate(rawCv);
+    await this.pubSub.publish('cvCreated', hydrated);
+    return hydrated;
   }
 
   @Mutation(() => CV)
   async updateCV(
     @Args('input') input: UpdateCVInput
   ) {
-    // check if skillIds are valid
     if(input.skillIds) {
       input.skillIds = input.skillIds.map(skillId => parseInt(skillId.toString(), 10));
       const skills = input.skillIds.map(skillId => this.skillService.findOne(skillId));
@@ -65,23 +69,28 @@ export class CvResolver {
     
     const updated = this.cvService.update(input.id, input);
     if (!updated) throw new Error('CV not found');
-    return this.hydrate(updated);
+    const hydrated = await this.hydrate(updated);
+
+    await this.pubSub.publish('cvUpdated', hydrated);
+    
+    return hydrated;
   }
 
   @Mutation(() => Boolean)
   async deleteCV(
-    @Args('id') id: number,
-    @Context() ctx: { db: any }
-  ) {
-    id = parseInt(id.toString(), 10);
-    return this.cvService.delete(id);
+    @Args('id') id: number  ) {
+      const deletedCv = await this.cvService.findOne(id);
+      if (!deletedCv) throw new Error('CV not found');
+      const success = await this.cvService.delete(id);
+      if (success) await this.pubSub.publish('cvDeleted', deletedCv);
+      return success;
+
   }
 
-  private hydrate(cv: CV): CV {
-    return {
-      ...cv,
-      user: this.userService.findOne(cv.userId),
-      skills: cv.skillIds.map(skillId => this.skillService.findOne(skillId)),
-    } as CV;
+  private async hydrate(cv: CV): Promise<CV> {
+    const user = await this.userService.findOne(cv.userId);
+    const skills = (await Promise.all(cv.skillIds.map(id => this.skillService.findOne(id)))).filter((skill): skill is Skill => skill !== undefined);
+    return { ...cv, user, skills };
   }
+  
 }
